passed the token from frontend
If we ae not logged in we cant see the goals now all request must be send along with the token

+++ frontend/src/components/GoalDetail.js
@@ -1,12 +1,17 @@
 import React from 'react'
-import { Box, Button, Center, Checkbox, Flex, HStack, Text} from '@chakra-ui/react'
+import { Box, Button, Center, Checkbox, Flex, HStack, Text, useAccordionItem} from '@chakra-ui/react'
 import { useGoalcontext } from '../hooks/useGoalcontext'
-
+import { useAuthContext } from '../hooks/useAuthContext'
 export default function GoalDetail(props) {
   const {dispatch}=useGoalcontext()
+  const {user}=useAuthContext()
   const deleteHandler=async()=>{
+    if(!user){return}
     const data=await fetch("/api/goals/"+props.goal._id,{
-      method:'DELETE'
+      method:'DELETE',
+      headers:{
+        'Authorization':`Bearer ${user.token}`
+      }
     })
     
     const jsondata=await data.json()
